datagenie/requirements/base.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LLM and AI
openai==1.3.0
langchain==0.0.340
langchain-openai==0.0.2
tiktoken==0.5.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0  # PostgreSQL async driver
aiomysql==0.2.0  # MySQL async driver
aiosqlite==0.19.0  # SQLite async driver

# Data Processing
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2  # Excel file support

# Visualization
plotly==5.17.0

# Frontend
gradio==4.8.0

# Caching
redis==5.0.1
aioredis==2.0.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.8.0
cryptography==41.0.7
email-validator==2.1.0

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
rich==13.7.0
